yarab

+++ src/main/java/Indexer/wordAttr.java
@@ -5,7 +5,7 @@ import java.util.*;
 public class wordAttr {
 
     int priority=0;
-    int TF=0;
+    double TF=0;
     String link;
     ArrayList<String> snippets = new ArrayList<>();
     ArrayList<String> tags=new ArrayList<>();
